+++ app/frontend/src/javascript/components/user/user-profile-form.tsx
@@ -93,7 +93,7 @@ export const UserProfileForm: React.FC<UserProfileFormProps> = ({ action, size, 
       });
       setValue('invoicing_profile_attributes.user_profile_custom_fields_attributes', userProfileCustomFields);
     }).catch(error => onError(error));
-    SettingAPI.query(['phone_required', 'address_required', 'external_id'])
+    SettingAPI.query(['phone_required', 'address_required', 'external_id', 'gender_required', 'birthday_required'])
       .then(settings => setFieldsSettings(settings))
       .catch(error => onError(error));
   }, []);
@@ -185,7 +185,7 @@ export const UserProfileForm: React.FC<UserProfileFormProps> = ({ action, size, 
       <div className="fields-group">
         <div className="personnal-data">
           <h4>{t('app.shared.user_profile_form.personal_data')}</h4>
-          <GenderInput register={register} disabled={isDisabled} required tooltip={t('app.shared.user_profile_form.used_for_statistics')} />
+          <GenderInput register={register} disabled={isDisabled} required={fieldsSettings.get('gender_required') === 'true'} tooltip={t('app.shared.user_profile_form.used_for_statistics')} />
           <div className="names">
             <FormInput id="profile_attributes.last_name"
                        register={register}
@@ -205,7 +205,7 @@ export const UserProfileForm: React.FC<UserProfileFormProps> = ({ action, size, 
                        register={register}
                        label={t('app.shared.user_profile_form.date_of_birth')}
                        disabled={isDisabled}
-                       rules={{ required: true }}
+                       rules={{ required: fieldsSettings.get('birthday_required') === 'true' }}
                        formState={formState}
                        type="date"
                        nullable />