+++ app/models/user.rb
@@ -115,8 +115,10 @@ class User < ApplicationRecord
   end
 
   def need_completion?
-    statistic_profile.gender.nil? || profile.first_name.blank? || profile.last_name.blank? || username.blank? ||
-      email.blank? || encrypted_password.blank? || group_id.nil? || statistic_profile.birthday.blank? ||
+    (Setting.get('gender_required') && statistic_profile.gender.blank?) ||
+      profile.first_name.blank? || profile.last_name.blank? || username.blank? ||
+      email.blank? || encrypted_password.blank? || group_id.nil? ||
+      (Setting.get('birthday_required') && statistic_profile.birthday.blank?) ||
       (Setting.get('phone_required') && profile.phone.blank?) ||
       (Setting.get('address_required') && invoicing_profile&.address&.address&.blank?)
   end